+++ coordinator/app/src/main/kotlin/lineth/coordinator/app/CoordinatorApp.kt
@@ -24,6 +24,7 @@ import lineth.coordinator.app.conflation.ConflationAppOrchestrator
 import lineth.coordinator.app.conflation.TracesClientFactory.createTracesClients
 import lineth.coordinator.app.conflation.TracesClients
 import lineth.coordinator.app.conflationbacktesting.ConflationBacktestingService
+import lineth.coordinator.clients.prover.ProverClientFactoryBuilder
 import lineth.coordinator.config.toJsonRpcRetry
 import lineth.coordinator.config.v2.CoordinatorConfig
 import lineth.coordinator.config.v2.DatabaseConfig
@@ -65,6 +66,7 @@ class CoordinatorApp(
   // behaves identically when no extension is supplied.
   extensionsFactory: CoordinatorExtensionFactory = CoordinatorExtensionFactory.NOOP,
   signerFactory: SignerFactory = DefaultSignerFactory,
+  proverClientFactoryBuilder: ProverClientFactoryBuilder = ProverClientFactoryBuilder.FILE_BASED,
 ) {
   private val log: Logger = LogManager.getLogger(this::class.java)
   private val vertx: Vertx =
@@ -268,6 +270,7 @@ class CoordinatorApp(
     l2EthClient = l2EthClientForConflation,
     zkStateClient = zkStateClient,
     tracesClients = tracesClients,
+    proverClientFactoryBuilder = proverClientFactoryBuilder,
   )
 
   private val l1FinalizationMonitorApp = L1FinalizationMonitorApp(