+++ coordinator/core-interfaces/src/main/kotlin/linea/clients/ProverFileNameProvider.kt
@@ -2,6 +2,12 @@ package linea.clients
 
 import linea.domain.ProofIndex
 
-interface ProverFileNameProvider<TProofIndex : ProofIndex> {
+interface ProverFileNameProvider<TProofIndex : ProofIndex> : ProverFileNameStartBlockNumberProvider {
   fun getFileName(proofIndex: TProofIndex): String
+}
+
+interface ProverFileNameStartBlockNumberProvider {
+  fun getFileNameStartBlockNumber(fileName: String): Long {
+    return fileName.substringBefore('-').toULong().toLong()
+  }
 }