+++ tracer/testing/src/main/java/net/consensys/linea/testing/ShomeiNode.java
@@ -10,7 +10,7 @@
  * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
  * specific language governing permissions and limitations under the License.
  *
- * SPDX-License-Identifier: MIT OR Apache-2.0
+ * SPDX-License-Identifier: Apache-2.0
  */
 
 package net.consensys.linea.testing;